:recycle: Refactor: Upload service refactor
Upload service refactor
