+++ resources/views/dashboard/search/index.blade.php
@@ -38,7 +38,7 @@
                         <td>{{ $post->id }}</td>
                         <td>
                             @if ($post->photo)
-                            <img src="{{ asset('storage/' . $post->photo->path) }}" height="60" width="90" alt="Photo">
+                            <img src="{{ Storage::url($post->photo->path) }}" height="60" width="90" alt="Photo">
                             @endif
                         </td>
                         <td>
