+++ resources/views/frontend/post/show.blade.php
@@ -22,7 +22,7 @@
             <div class="thumbnail">
                 <img class="lazyload"
                     src="data:image/gif;base64,R0lGODlhAgABAIAAAP///wAAACH5BAEAAAEALAAAAAACAAEAAAICTAoAOw=="
-                    data-src="{{ asset('storage/' . $post->photo->path) }}" alt="{{ $post->title }}">
+                    data-src="{{ Storage::url($post->photo->path) }}" alt="{{ $post->title }}">
             </div>
             @if ($post->photo_source)
             <p class="photo-source">{{ $post->photo_source }}</p>
