+++ resources/views/index.blade.php
@@ -26,7 +26,7 @@
             <a href="{{ route('post.show', [$featured->slug]) }}">
                 <img class="lazyload" 
                     src="data:image/gif;base64,R0lGODlhAgABAIAAAP///wAAACH5BAEAAAEALAAAAAACAAEAAAICTAoAOw=="
-                    data-src="{{ asset('storage/' . $featured->photo->path) }}" alt="{{ $featured->title }}">
+                    data-src="{{ Storage::url($featured->photo->path) }}" alt="{{ $featured->title }}">
                 <div class="image-overlay"></div>
             </a>
         </div>
@@ -47,7 +47,7 @@
                 <a href="{{ route('post.show', [$post_item->slug]) }}">
                     <img class="lazyload"
                         src="data:image/gif;base64,R0lGODlhAgABAIAAAP///wAAACH5BAEAAAEALAAAAAACAAEAAAICTAoAOw=="
-                        data-src="{{ asset('storage/' . $post_item->photo->path) }}" alt="{{ $post_item->title }}">
+                        data-src="{{ Storage::url($post_item->photo->path) }}" alt="{{ $post_item->title }}">
                     <div class="image-overlay"></div>
                 </a>
             </div>
@@ -133,7 +133,7 @@
                     <a href="{{ route('post.show', [$post_item->slug]) }}">
                         <img class="lazyload" 
                             src="data:image/gif;base64,R0lGODlhAgABAIAAAP///wAAACH5BAEAAAEALAAAAAACAAEAAAICTAoAOw=="
-                            data-src="{{ asset('storage/' . $post_item->photo->path) }}" alt="{{ $post_item->title }}">
+                            data-src="{{ Storage::url($post_item->photo->path) }}" alt="{{ $post_item->title }}">
                         <div class="image-overlay"></div>
                     </a>
                 </div>
